[CVSRP-4593] create EmployeeTypeUtil WIP
